fix warnings
svn path=/trunk/koffice/; revision=139281

+++ karbon/ui/vmdlg_solidfill.cc
@@ -125,7 +125,7 @@ void VMDlgSolidFill::slotUpdateFromRGBSpinBoxes()
 	mSelector->blockSignals( true );
 	mColorSelector->setValues( h, s );
 	slotHSChanged( h, s );
-	mSelector->setValue( ( float( mValue->value() ) / 255.0 ) * 99.0 );
+	mSelector->setValue( static_cast<int>(( float( mValue->value() ) / 255.0 ) * 99.0) );
 	mSelector->blockSignals( false );
 
 	mHue->blockSignals( false );
@@ -141,7 +141,7 @@ void VMDlgSolidFill::slotUpdateFromHSVSpinBoxes()
 
 	// update gradient selector
 	mSelector->blockSignals( true );
-	mSelector->setValue( ( float( mValue->value() ) / 255.0 ) * 99.0 );
+	mSelector->setValue( static_cast<int>(( float( mValue->value() ) / 255.0 ) * 99.0) );
 	mSelector->blockSignals( false );
 
 	// set RGB
@@ -190,7 +190,7 @@ void VMDlgSolidFill::slotHSChanged( int h, int s )
 void VMDlgSolidFill::slotVChanged( int newVal )
 {
 	//QColor color( mHue->value(), mSaturation->value(), newVal, QColor::Hsv );
-	mValue->setValue( float( newVal ) / 99.0 * 255.0 );
+	mValue->setValue( static_cast<int>(float( newVal ) / 99.0 * 255.0) );
 }
 
 #include "vmdlg_solidfill.moc"
